.NET runtime: Fan.Sys.Float => Fan.Sys.Double

+++ src/nfan/fan/sys/Buf.cs
@@ -234,9 +234,9 @@ namespace Fan.Sys
 
     public Buf writeI8(Int x) { m_out.writeI8(x); return this; }
 
-    public Buf writeF4(Float x) { m_out.writeF4(x); return this; }
+    public Buf writeF4(Double x) { m_out.writeF4(x); return this; }
 
-    public Buf writeF8(Float x) { m_out.writeF8(x); return this; }
+    public Buf writeF8(Double x) { m_out.writeF8(x); return this; }
 
     public Buf writeDecimal(Decimal x) { m_out.writeDecimal(x); return this; }
 
@@ -290,9 +290,9 @@ namespace Fan.Sys
 
     public Int readS8() { return m_in.readS8(); }
 
-    public Float readF4() { return m_in.readF4(); }
+    public Double readF4() { return m_in.readF4(); }
 
-    public Float readF8() { return m_in.readF8(); }
+    public Double readF8() { return m_in.readF8(); }
 
     public Decimal readDecimal() { return m_in.readDecimal(); }
